Remove old updates that spanned several schema changes and require that database be current before upgrading to 6.x. This is needed because the CCK database has dramatically changed several times, including changing to and then away from table names that were later usurped by core. The 6.x structure finishes moving completely away from using table names prefixed with 'node_' to avoid future upgrade problems as core incorporates more parts of CCK.
This means 4.7 to 6.x upgrades will require an intermediate update to 5.x, and 5.x to 6.x upgrades will require that the latest code be used to upgrade the database in 5.x before upgrading to 6.x.

This won't be popular, but maintaining all those old updates to make sure that nothing breaks no matter what state the database is in has become an impossible task. Hopefully once everyone is stablized using the 6.x structure we can stick to automatic updates in the future.