Re-name fields to use Views 1 naming like field_text_value instead of field_text to make updates easier, and alter content.views_convert.inc to match.
Add default filters for all fields, and update fields to switch from string or numeric filters to 'in' operators if there are lists of allowed values.

Add helper function to return Views table name to make it easier to change the tablename in the future, like if we find a way to avoid the need to alias each field as its own table.